% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define_vars.R
\name{define_vars_}
\alias{define_vars_}
\title{This is a hack, still in progress}
\usage{
define_vars_(varnames)
}
\arguments{
\item{varname}{Character vector of variable names to create definitions for}
}
\description{
This is a hack, still in progress
}

