% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NSE.R
\docType{data}
\name{dplyr_NSE}
\alias{dplyr_NSE}
\title{For the code that created this, see the examples}
\format{\preformatted{ chr [1:20] "arrange" "count" "data_frame" "distinct" "do" ...
}}
\usage{
dplyr_NSE
}
\description{
Character vector of functions in dplyr
}
\examples{
library("dplyr")
funcs <- ls("package:dplyr")
ret <- funcs[grepl(".*_$", funcs)]
dput(gsub("_$", "", ret))
}
\keyword{datasets}

