% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reprex_prefix.R
\name{reprex_prefix}
\alias{reprex_prefix}
\title{Create a reproducible prefix for some code that loads the necessary
packages and defines variables}
\usage{
reprex_prefix(code)
}
\arguments{
\item{code}{Code as a character vector}
}
\description{
Create a reproducible prefix for some code that loads the necessary
packages and defines variables
}

