% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vars_used.R
\name{vars_used}
\alias{vars_used}
\title{Retrieve variables and packages used in an expression}
\usage{
vars_used(expr)
}
\arguments{
\item{expr}{An expression object}
}
\description{
Retrieve variables and packages used in an expression
}

